Move most methods from main.go to gosplit.go
